NCBC-2929: LoggingMeter emit_interval_s wrong unit
Motivation
----------
This commit makes the LoggingMeter use secconds for the emit_interval_s
field.

Reviewed-on: http://review.couchbase.org/c/couchbase-net-client/+/158033
Tested-by: Build Bot <build@couchbase.com>

+++ src/Couchbase/Core/Diagnostics/Metrics/LoggingMeterReport.cs
@@ -13,7 +13,7 @@ namespace Couchbase.Core.Diagnostics.Metrics
         public Meta? meta { get; set; }
         public Operations? operations { get; set; }
 
-        public static LoggingMeterReport Generate(ReadOnlyDictionary<string, IMetricsRoot?> histograms, double interval)
+        public static LoggingMeterReport Generate(ReadOnlyDictionary<string, IMetricsRoot?> histograms, uint interval)
         {
             var report = new LoggingMeterReport
             {
@@ -82,7 +82,7 @@ namespace Couchbase.Core.Diagnostics.Metrics
 
     internal class Meta
     {
-        public double emit_interval_s { get; set; }
+        public uint emit_interval_s { get; set; }
     }
 
     internal class PercentilesUs
